Development work on HTTP/2
This is based on Cory Benfield's initial work in
https://github.com/KeepSafe/aiohttp/pull/882 -- essentially, it splices
his code directly into the respective classes and tries to remove as
much duplication as possible, while adding ugly if/else switches.

The resulting server is capable of switching between HTTP/1.1 and
(limited) HTTP/2 communication based on the existence of an HTTP/2
PRIamble in the first request.

This is NOT PRODUCTION READY and SHOULD NOT BE MERGED. It's just
copy-paste code, intended to start showing the seams where we can
extract new interfaces and refactor as necessary.